% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_load_installr_on_startup_menu}
\alias{add_load_installr_on_startup_menu}
\title{Add menu item for having installr load on startup}
\usage{
add_load_installr_on_startup_menu(...)
}
\arguments{
  \item{...}{not used. (but good for future backward
  compatibility)}
}
\description{
Add menu item for having installr load on startup
}

