\name{add_remove_installr_from_startup_menu}
\alias{add_remove_installr_from_startup_menu}
\title{Add menu item for having installr NOT load on startup}
\usage{
  add_remove_installr_from_startup_menu(...)
}
\arguments{
  \item{...}{not used. (but good for future backward
  compatibility)}
}
\description{
  Add menu item for having installr NOT load on startup
}

