% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{add_to_.First_in_Rprofile.site}
\alias{add_to_.First_in_Rprofile.site}
\title{Add a code line to Rprofile.site .First}
\usage{
add_to_.First_in_Rprofile.site(code, indent = "\\t", ...)
}
\arguments{
\item{code}{A character scalar with code to add at the beginning of the
.First function in Rprofile.site}

\item{indent}{a character scalar indicating the text to be added before
code. Default is a tab.}

\item{...}{not used.}
}
\description{
Goes through Rprofile.site text, finds the .First function - and add
a line of code to the beginning of it.
}
\examples{
\dontrun{
is_in_.First_in_Rprofile.site("suppressMessages(library(installr))") # FALSE
add_to_.First_in_Rprofile.site("suppressMessages(library(installr))")
is_in_.First_in_Rprofile.site("suppressMessages(library(installr))") # TRUE
remove_from_.First_in_Rprofile.site("suppressMessages(library(installr))")
is_in_.First_in_Rprofile.site("suppressMessages(library(installr))") # FALSE
# this would still leave .First
}
}
\references{
\url{https://stackoverflow.com/questions/1395301/how-to-get-r-to-recognize-your-working-directory-as-its-working-directory}
\url{https://stackoverflow.com/questions/1189759/expert-r-users-whats-in-your-rprofile}
\url{http://www.noamross.net/archives/2012-11-02-rprofile/}
\url{https://www.statmethods.net/interface/customizing.html}
}
