% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create.global.library.R
\name{create.global.library}
\alias{create.global.library}
\title{Creates a global library folder}
\usage{
create.global.library(global_library_folder)
}
\arguments{
  \item{global_library_folder}{the path of the new global
  library folder to create.  If missing, will be set to
  R_path/R/library. (for example: "C:/Program
  Files/R/library")}
}
\value{
TRUE/FALSE if we created a new folder or not.
}
\description{
Creates a global library folder (above the folder R is
currently installed in)
}
\examples{
\dontrun{
create.global.library()
}
}

