% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateR.R
\name{get_latest_r_version}
\alias{get_latest_r_version}
\title{Fetches latest R version from CRAN}
\usage{
get_latest_r_version(cran_mirror = "https://cran.rstudio.com/")
}
\arguments{
\item{cran_mirror}{The CRAN mirror to use}
}
\description{
Returns the latest version of R available on CRAN as an R system version object.
}
\references{
\url{https://cran.r-project.org/bin/windows/base/}
}
\seealso{
\link{getRversion}
}
