% Generated by roxygen2 (4.0.2): do not edit by hand
\name{install.Cygwin}
\alias{install.Cygwin}
\alias{install.cygwin}
\title{Downloads and installs Cygwin for Windows}
\usage{
install.Cygwin(bit = if (Sys.info()["machine"] == "x86-64") "64" else "32",
  ...)
}
\arguments{
  \item{bit}{Specify whether to install the 32 or 64 bit
  version.  The default is to match the version of Windows
  running on your machine.}

  \item{...}{extra parameters to pass to
  \link{install.URL}}
}
\value{
TRUE/FALSE - was the installation successful or not.
}
\description{
Allows the user to download and install the latest version
of Cygwin for Windows.
}
\details{
Cygwin is a collection of tools which provide a Linux look
and feel environment for Windows.
}
\examples{
\dontrun{
install.Cygwin() # installs the latest version of Cygwin
}
}
\references{
\itemize{ \item Cygwin homepage: \url{http://cygwin.com/} }
}

