% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install.git}
\alias{install.git}
\title{Downloads and installs git and git-gui for windows}
\usage{
install.git(URL = "https://git-scm.com/download/win", version = 64, ...)
}
\arguments{
\item{URL}{the URL of the git download page.}

\item{version}{numeric - either 32 or 64 (default)}

\item{...}{extra parameters to pass to \link{install.URL}}
}
\value{
TRUE/FALSE - was the installation successful or not.
}
\description{
Allows the user to downloads and install the latest version of git for Windows.
}
\details{
Git is a distributed revision control and source code management system with an emphasis on speed.
}
\examples{
\dontrun{
install.git() # installs the latest version of git
}
}
\references{
git homepage: \url{https://git-scm.com/}
git download page: \url{https://git-scm.com/download/win}
}
