% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installNodejs.R
\name{install.nodejs}
\alias{install.nodejs}
\title{Downloads and installs nodejs LTS or Current}
\usage{

  install.nodejs(page_with_download_url = "https://nodejs.org/en/download/",
  version_number = "LTS", ...)
}
\arguments{
\item{page_with_download_url}{a link to the list of download links for Nodejs}

\item{version_number}{Either LTS or Current. Version LTS will lead to download of v6.11.X}

\item{...}{extra parameters to pass to \link{install.URL}}
}
\value{
TRUE/FALSE - was the installation successful or not.
}
\description{
Downloads and installs the latest version of nodejs LTS or Current for Windows.
}
\details{
Nodejs is a programming language which has two versions under active development. 
Make sure you know which version is required for the code you have to run, or alternatively, make sure you are developing code that is fit for your chosen version of Nodejs.
}
\examples{
\dontrun{
install.nodejs() 
install.nodejs(,"Current")
install.nodejs(,"LTS")
}
}
\author{
Tal Galili and A. Jonathan R. Godfrey and Chanyub Park
}
