% Generated by roxygen2 (4.0.2): do not edit by hand
\name{install.OpenSSH}
\alias{install.OpenSSH}
\title{Downloads and installs OpenSSH for Windows}
\usage{
install.OpenSSH(...)
}
\arguments{
  \item{...}{extra parameters to pass to
  \link{install.URL}}
}
\value{
TRUE/FALSE - was the installation successful or not.
}
\description{
Allows the user to download and install OpenSSH (version
6.7p1-1-v1) for Windows.
}
\details{
SSH (\strong{s}ecure \strong{sh}ell) is a protocol for
securely connecting to a remote device.  OpenSSH is a free
version of SSH, developed by the OpenBSD Project.  This
installer for Windows is provided by mls-software.com.
OpenSSH can create an SSH connection from the command
prompt.  Cygwin (\link{install.Cygwin}) provides similar
capabilities (among others), but a full installation of
Cygwin is much larger than OpenSSH.
}
\examples{
\dontrun{
install.OpenSSH() # installs version 6.7p1-1-v1 of OpenSSH
}
}
\references{
\itemize{ \item SSH page on Wikipedia:
\url{http://en.wikipedia.org/wiki/Secure_Shell} \item
OpenSSH homepage: \url{http://www.openssh.com/} \item
OpenSSH page on mls-software.com:
\url{http://www.mls-software.com/opensshd.html} \item
install.Cygwin: \link{install.Cygwin} }
}

