% Generated by roxygen2 (4.0.1): do not edit by hand
\name{install.Texmaker}
\alias{install.Texmaker}
\alias{install.texmaker}
\title{Downloads and installs Texmaker for windows}
\usage{

  install.Texmaker(URL = "http://www.xm1math.net/texmaker/texmakerwin32_install.exe",
  ...)
}
\arguments{
  \item{URL}{the URL of the GitHub download page.}

  \item{...}{extra parameters to pass to
  \link{install.URL}}
}
\value{
logical - was the installation successful or not.
}
\description{
Allows the user to downloads and install the latest version
of Texmaker for Windows.
}
\details{
Texmaker is a free, modern and cross-platform LaTeX editor
for linux, macosx and windows systems that integrates many
tools needed to develop documents with LaTeX, in just one
application.
}
\examples{
\dontrun{
install.Texmaker() # installs the latest version of Texmaker for windows
}
}
\references{
\itemize{ \item Texmaker homepage:
\url{http://www.xm1math.net/texmaker/} }
}

