% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_from_list.R
\name{install_from_list}
\alias{install_from_list}
\title{Install then load libraries from vector}
\usage{
install_from_list(libs, load = TRUE)
}
\arguments{
\item{libs}{A vector}

\item{load}{TRUE/FALSE}
}
\value{
none
}
\description{
Install then load libraries from vector
}
\examples{
#install_from_list(c("caret"), TRUE)
}
