% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_from_script.R
\name{install_from_script}
\alias{install_from_script}
\title{Install then load libraries from script file(s)}
\usage{
install_from_script(files, load = TRUE)
}
\arguments{
\item{files}{A vector with file names}

\item{load}{TRUE/FALSE}
}
\value{
none
}
\description{
Install then load libraries from script file(s)
}
\examples{
#install_from_script(c("file1.R", "file2.R"), TRUE)
#install_from_script("file.R", FALSE)
}
