% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_from_yml.R
\name{install_from_yml}
\alias{install_from_yml}
\title{Install then load libraries from YAML file}
\usage{
install_from_yml(file = "deps.yml")
}
\arguments{
\item{file}{YAML file path}
}
\value{
none
}
\description{
Install then load libraries from YAML file
}
\examples{
#install_from_yml("deps.yml")
}
