% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstallLibs.R
\name{InstallLibs}
\alias{InstallLibs}
\title{Install then load libraries from vector}
\usage{
InstallLibs(libs, load = TRUE)
}
\arguments{
\item{libs}{A vector}

\item{load}{TRUE/FALSE}
}
\value{
none
}
\description{
Install then load libraries from vector
}
\examples{
InstallLibs(c("caret"), TRUE)
}
