% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstallLibsScript.R
\name{InstallLibsScript}
\alias{InstallLibsScript}
\title{Install then load libraries from script file(s)}
\usage{
InstallLibsScript(files, load = TRUE)
}
\arguments{
\item{files}{A vector with file names}

\item{load}{TRUE/FALSE}
}
\value{
none
}
\description{
Install then load libraries from script file(s)
}
\examples{
InstallLibsScript(c("file1.R", "file2.R"), TRUE)
InstallLibsScript("file.R", FALSE)
}
