% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstallLibsYML.R
\name{InstallLibsYML}
\alias{InstallLibsYML}
\title{Install then load libraries from YAML file}
\usage{
InstallLibsYML(file = "deps.yml")
}
\arguments{
\item{file}{YAML file path}
}
\value{
none
}
\description{
Install then load libraries from YAML file
}
\examples{
InstallLibsYML("deps.yml")
}
