% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_functions.R
\name{iplookup}
\alias{iplookup}
\title{Geolocate IP address using IP2Location.io in R}
\source{
\url{https://api.ip2location.io}
}
\usage{
iplookup(ip, key)
}
\arguments{
\item{ip}{IPv4 or IPv6 address}

\item{key}{Your IP2Location.io API Key}
}
\value{
API result from IP2Location.io API. Visit the Response Format section in \url{https://www.ip2location.io/ip2location-documentation} to learn more on the fields available in the result.
}
\description{
This R function uses the IP2Location.io geolocation API to look up for an IP address for a set of enriched information. This includes country, region, city, latitude and longtitude, zipcode and ASN. For more information about the API, kindly visit the documentation at \url{https://www.ip2location.io/ip2location-documentation}.

The function require rjson.
}
\examples{
\dontrun{
iplookup("8.8.8.8", "YOUR_API_KEY")
}
}
\seealso{
\link{freegeoip}, \link{myip}, \link{cranometer}
}
\author{
IP2Location
}
