% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.x.R
\name{is.exe.installed}
\alias{is.exe.installed}
\title{Checks if some .exe is available in on the Windows machine search PATH}
\usage{
is.exe.installed(exe_file)
}
\arguments{
\item{exe_file}{a character with the name of the executable to be looked for}
}
\value{
A boolean vector indicating the existence of each program's availability  on the system.
}
\description{
Checks the existence of an .exe extention in the search path for executable files
}
\examples{
\dontrun{
is.exe.installed(c("zip.exe", "Rgui.exe", "blablabla")) # [1]  TRUE  TRUE FALSE
is.exe.installed("7z") 
}
}
