\name{is.Rgui}
\alias{is.Rgui}
\title{Checks if the R session is running within Rgui (Windows OS)}
\usage{
  is.Rgui()
}
\value{
  Returns TRUE/FALSE if the R session is running within
  Rgui or not.
}
\description{
  Returns TRUE/FALSE if the R session is running within
  Rgui or not.
}
\details{
  This function is used in order to check if a GUI can be
  added to the session or not.
}
\examples{
\dontrun{
is.Rgui()
}
}
\seealso{
  \link{is.RStudio}, \link{is.windows}
}

