% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is.x.R
\name{is.RStudio}
\alias{is.RStudio}
\title{Checks if the R session is running within RStudio}
\usage{
is.RStudio(...)
}
\arguments{
  \item{...}{none are available.}
}
\value{
Returns TRUE/FALSE if the R session is running within
RStudio or not.
}
\description{
Returns TRUE/FALSE if the R session is running within
RStudio or not.
}
\details{
This function is used in order to check if a GUI can be
added to the session or not.
}
\examples{
\dontrun{
is.RStudio()
}
}

