\name{is.windows}
\alias{is.windows}
\title{Checks if the running OS is windows}
\usage{
  is.windows(...)
}
\arguments{
  \item{...}{non are available.}
}
\value{
  Returns TRUE/FALSE if the R session is running on Windows
  or not.#'
}
\description{
  Returns TRUE/FALSE if the R session is running on Windows
  or not.
}
\details{
  Since the 'installr' package is written mainly for
  Windows - it is a good thing to know if it is run on
  Windows or not.
}
\examples{
is.windows() # returns TRUE on my machine.
}

