% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.x.R
\name{is.x64}
\alias{is.x64}
\title{Checks if the running OS is x64}
\usage{
is.x64(...)
}
\arguments{
\item{...}{none are available.}
}
\value{
Returns TRUE/FALSE if the R session is running on Windows 64-bit or not.
}
\description{
Returns TRUE/FALSE if the R session is running on Windows 64-bit or not.
}
\details{
This function is run when the 'installr' package is first loaded in order to check if the current running OS is Windows 64-bit.
If you are running a different OS, then the installr package (at its current form) does not have much to offer you.
}
\examples{
is.x64() # returns TRUE on my machine.
}
