% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kill_pid.R
\name{kill_all_Rscript_s}
\alias{kill_all_Rscript_s}
\title{kill (i.e.: stop) all running "Rscript" processes}
\usage{
kill_all_Rscript_s(s = 0, m = 0, h = 0, ...)
}
\arguments{
\item{s}{numeric. number of seconds to wait before killing the processes}

\item{m}{numeric. number of minutes to wait before killing the processes}

\item{h}{numeric. number of hours to wait before killing the processes}

\item{...}{not used.}
}
\value{
an integer vector with the process ID (pid) of the "Rscript" processes.
}
\description{
kill (i.e.: stop) all running "Rscript" processes 
based on their process ID (pid)
}
\examples{
\dontrun{
# create several running processes of Rscript (to shitdown)
system("Rscript -e repeat{2+2}", wait = FALSE) # this process should be stuck   
system("Rscript -e repeat{2+2}", wait = FALSE) # this process should be stuck   
# here are there pid numbers:
get_Rscript_PID() 
# let's kill them:
kill_all_Rscript_s()
# they are gone...
get_Rscript_PID() # we no longer have Rscripts running
}
}
\references{
tasklist details from microsoft homepage: \url{https://docs.microsoft.com/en-us/previous-versions/windows/it-pro/windows-xp/bb491010(v=technet.10)}
pskill details from microsoft homepage: \url{https://docs.microsoft.com/en-us/sysinternals/downloads/pskill}
}
\seealso{
\code{\link{get_tasklist}}, \code{\link{get_Rscript_PID}}, \code{\link{get_pid}},
\code{\link{kill_pid}}, \code{\link{kill_all_Rscript_s}}, \link[tools]{pskill}
}
