% Generated by roxygen2 (4.0.2): do not edit by hand
\name{R_version_in_a_folder}
\alias{R_version_in_a_folder}
\title{Get the version of the R installed in a folder}
\usage{
R_version_in_a_folder(folder)
}
\arguments{
  \item{folder}{The folder for which we wish to know the R
  version.}
}
\value{
Returns a character vector of the R version (or NA, if this
is not an R installation folder)
}
\description{
Get the version of the R installed in a folder based on the
structure of the filename README.R-... (where ... is a
version number for R). This function helps detect the
version number of an R installation even if the name of the
folder is not standard. If multiple versions were
installed, overwriting each other, the most recent is
selected.
}
\examples{
\dontrun{
R_version_in_a_folder(R.home())
# returns the version of the current R installation
}
}
\seealso{
\link{get.installed.R.folders}
}

