% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rename_r_to_R}
\alias{rename_r_to_R}
\title{Rename files' extensions in a folder from .r to .R}
\usage{
rename_r_to_R(subdir, ...)
}
\arguments{
  \item{subdir}{sub folder from the current working
  directoryl in which the files should be changed}

  \item{...}{not used.}
}
\value{
(integer) the number of files changed
}
\description{
Rename files' extensions in a folder from .r to .R.
}
\details{
This came after a discussion with Hadley, JJ, and Martin
leading to the realization that since we are using the R
language (and not the r language), the standard is to use
.R files instead of .r
}
\examples{
\dontrun{
rename_r_to_R("R") # fixing the file ending inside a package directory
}
}

