% Generated by roxygen2 (4.0.2): do not edit by hand
\name{restart_RGui}
\alias{restart_RGui}
\title{Restart RGui from RGui}
\usage{
restart_RGui(...)
}
\arguments{
  \item{...}{passed to q()}
}
\value{
q(...)
}
\description{
Start a new RGui session and then quites the current one.

This is a Windows only function.
}
\examples{
\dontrun{
restart_RGui()
}
}

