% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/updateR.R
\name{turn.version.to.number1}
\alias{turn.version.to.number1}
\title{Turns version to number (for 1 value only)}
\usage{
turn.version.to.number1(version_with_dots)
}
\arguments{
  \item{version_with_dots}{A character value - of the
  version of R (for example 2.15.2)}
}
\value{
A "number" representation of the version (for example:
2015002)
}
\description{
Turns version to number (for 1 value only)
}
\examples{
\dontrun{
turn.version.to.number1("2.15.2")
turn.version.to.number1("3.0.1")
}
}
\seealso{
\link{turn.version.to.number}
}

