% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx2csv.R
\name{xlsx2csv}
\alias{xlsx2csv}
\title{Converts xls(x) to csv using VB}
\source{
This is based on the code from plang's answer here:
\url{http://stackoverflow.com/questions/1858195/convert-xls-to-csv-on-command-line}
}
\usage{
xlsx2csv(xlsx, csv, path, ...)
}
\arguments{
\item{xlsx}{the (character) name of the xlsx (or xls) file to convert.
if xlsx has a full path, it will override the path parameter.}

\item{csv}{the (character) name of the csv file to convert to (default will be the name of the xlsx file)}

\item{path}{the path for the files (default is the working diractory).}

\item{...}{ignored.}
}
\examples{
\dontrun{

xlsx2csv("c:/some_file.xlsx")

}
}

