% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kill_pid.R
\name{get_Rscript_PID}
\alias{get_Rscript_PID}
\title{Get the running "Rscript" processes PID}
\usage{
get_Rscript_PID(...)
}
\arguments{
\item{...}{not used.}
}
\value{
an integer vector with the process ID (pid) of the "Rscript" processes.
}
\description{
Returns a vector with the process ID (pid) of the "Rscript" 
processes which are currently running.
}
\examples{
\dontrun{
# create several running processes of Rscript (to shitdown)
system("Rscript -e repeat{2+2}", wait = FALSE) # this process should be stuck   
system("Rscript -e repeat{2+2}", wait = FALSE) # this process should be stuck   
# here are there pid numbers:
get_Rscript_PID() 
# let's kill them:
kill_all_Rscript_s()
# they are gone...
get_Rscript_PID() # we no longer have Rscripts running
}
}
\references{
tasklist details from microsoft homepage: \url{https://learn.microsoft.com/en-us/previous-versions/windows/it-pro/windows-xp/bb491010(v=technet.10)}
}
\seealso{
\code{\link{get_tasklist}}, \code{\link{get_Rscript_PID}}, \code{\link{get_pid}},
\code{\link{kill_pid}}, \code{\link{kill_all_Rscript_s}}, \link[tools]{pskill}
}
