% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/install.R
\name{install.FFmpeg}
\alias{install.FFmpeg}
\alias{install.ffmpeg}
\title{Downloads and installs FFmpeg for windows}
\usage{
install.FFmpeg(page_with_download_url = "http://ffmpeg.zeranoe.com/builds/",
  ...)
}
\arguments{
  \item{page_with_download_url}{the URL of the FFmpeg
  download page.} \item{...}{extra parameters to pass to
  \link{install.URL}}
}
\description{
Allows the user to downloads the latest version of FFmpeg
for Windows. IMPORTANT NOTE: The user (YOU) are responsible
for unpacking the 7zip file into the relevant directory.
All that this function does is to download the 7zip file and
"run" it.
}
\details{
FFmpeg is a complete, cross-platform solution to record,
convert and stream audio and video. It includes libavcodec -
the leading audio/video codec library. See the documentation
for a complete feature list and the Changelog for recent
changes. This function downloads current releases and NOT
the Development Snapshots. This function is useful for
saveVideo() in the {animation} package.
}
\examples{
\dontrun{
install.FFmpeg() # installs the latest version of FFmpeg
}
}
\references{
\itemize{ \item FFmpeg homepage: \url{http://FFmpeg.org/} }
}

