% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installOpenJdk.R
\name{install.java}
\alias{install.java}
\alias{install.Java}
\alias{install.jdk}
\alias{install.Jdk}
\alias{install.openjdk}
\alias{install.OpenJdk}
\alias{install.java}
\alias{install.Java}
\alias{install.jdk}
\alias{install.Jdk}
\alias{install.openjdk}
\alias{install.OpenJdk}
\title{Install Java - downloads and set path openjdk}
\usage{
install.java(version = 10,
  page_with_download_url = "http://jdk.java.net/java-se-ri/",
  path = "C:/java")

install.Java(version = 10,
  page_with_download_url = "http://jdk.java.net/java-se-ri/",
  path = "C:/java")

install.jdk(version = 10,
  page_with_download_url = "http://jdk.java.net/java-se-ri/",
  path = "C:/java")

install.Jdk(version = 10,
  page_with_download_url = "http://jdk.java.net/java-se-ri/",
  path = "C:/java")

install.openjdk(version = 10,
  page_with_download_url = "http://jdk.java.net/java-se-ri/",
  path = "C:/java")

install.OpenJdk(version = 10,
  page_with_download_url = "http://jdk.java.net/java-se-ri/",
  path = "C:/java")
}
\arguments{
\item{version}{9 or 10 is passible. Default is 10.}

\item{page_with_download_url}{where to download. Default is \url{http://jdk.java.net/java-se-ri/}}

\item{path}{where to set java. Defulat path is C:/java}
}
\value{
TRUE/FALSE - was the installation successful or not.
}
\description{
Downloads and set path the latest version of openjdk for Windows.
}
\details{
install openjdk 9 or 10 version for windows.
}
\examples{
\dontrun{
install.java() 
install.java(version = 10)
install.java(10)
}
}
\author{
Chan-Yub Park And Tal Galili
}
