% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateR.R
\name{install.R}
\alias{install.R}
\title{Downloads and installs the latest R version}
\usage{

  install.R(page_with_download_url = "https://cran.rstudio.com/bin/windows/base/",
  pat = "R-[0-9.]+-win.exe", to_checkMD5sums = TRUE,
  keep_install_file = FALSE, download_dir = tempdir(), silent = FALSE,
  ...)
}
\arguments{
\item{page_with_download_url}{URL from which the latest stable version of R can be downloaded from.}

\item{pat}{the pattern of R .exe file to download}

\item{to_checkMD5sums}{Should we check that the new R installation has the files we expect it to (by checking the MD5 sums)? default is TRUE.  It assumes that the R which was isntalled is the latest R version.}

\item{keep_install_file}{If TRUE - the installer file will not be erased after it is downloaded and run.}

\item{download_dir}{A character of the directory into which to download the file. (default is \link{tempdir}())}

\item{silent}{If TRUE - enables silent installation mode.}

\item{...}{extra parameters to pass to \link{install.URL}}
}
\value{
TRUE/FALSE - was the installation of R successful or not.
}
\description{
Fetches the latest (not development!) R version
}
\details{
If you are not sure if you need to update R or not, 
It is better to use updateR for updating R, since it includes more options.
But in case you wish to only install R, with no other steps taken (for example, taking care of your old packages), then you can use install.R()

See the \link{install.Rdevel} function for installing the latest R-devel version.
}
\examples{
\dontrun{
install.R() 
}
}
\references{
\url{https://cran.rstudio.com/bin/windows/base/}
}
\seealso{
\link{uninstall.R}, \link{install.Rdevel}, \link{updateR}, \link{system}
}

