% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/install.R
\name{install.RStudio}
\alias{install.RStudio}
\alias{install.rstudio}
\title{Downloads and installs RStudio for windows}
\usage{

  install.RStudio(page_with_download_url = "http://www.rstudio.com/ide/download/desktop",
  ...)
}
\arguments{
  \item{page_with_download_url}{the URL of the RStudio
  download page.}

  \item{...}{extra parameters to pass to
  \link{install.URL}}
}
\value{
TRUE/FALSE - was the installation successful or not.
}
\description{
Allows the user to downloads and install the latest version
of RStudio for Windows.
}
\details{
RStudio is a free and open source integrated development
environment (IDE) for R, a programming language for
statistical computing and graphics.
}
\examples{
### devtools::source_url
\dontrun{
install.RStudio() # installs the latest version of RStudio
}
}
\references{
\itemize{ \item RStudio homepage:
\url{http://www.rstudio.com/} }
}

