% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installr-package.R
\docType{package}
\name{installr-package}
\alias{installr-package}
\title{Using R to Install Stuff (Such As: R, Rtools, RStudio, Git, and More!)}
\description{
Using R to Install Stuff (Such As: R, Rtools, RStudio, Git, and More!)
}
\keyword{installr}
