% Generated by roxygen2 (4.0.1): do not edit by hand
\name{load_installr_on_startup}
\alias{load_installr_on_startup}
\title{Have the installr package load on startup}
\usage{
load_installr_on_startup(...)
}
\arguments{
  \item{...}{not used. (but good for future backward
  compatibility)}
}
\value{
invisible(NULL)
}
\description{
Have the installr package load on startup
}
\examples{
\dontrun{
load_installr_on_startup()
}
}
\references{
\url{http://stackoverflow.com/questions/1395301/how-to-get-r-to-recognize-your-working-directory-as-its-working-directory}
\url{http://stackoverflow.com/questions/1189759/expert-r-users-whats-in-your-rprofile}
\url{http://www.noamross.net/blog/2012/11/2/rprofile.html}
\url{http://www.statmethods.net/interface/customizing.html}
}

