% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_functions.R
\name{myip}
\alias{myip}
\title{What is my IP}
\source{
\url{https://api.ipify.org}
}
\usage{
myip(...)
}
\arguments{
\item{...}{not in use}
}
\value{
your current ip (character string)
}
\description{
Retrieving your public IP via \url{https://api.ipify.org}.
(old solution used: http://api.exip.org/
based on http://stackoverflow.com/questions/3097589/getting-my-public-ip-via-api)
}
\examples{
\dontrun{
myip() # "37.132.25.15"
}
}
\seealso{
\link{freegeoip}, \link{myip}, \link{cranometer}
}
