% Generated by roxygen2 (4.0.2): do not edit by hand
\name{system.PATH}
\alias{system.PATH}
\title{Returns the search path for executable files}
\usage{
system.PATH()
}
\value{
A character vector with the search path for executable
files
}
\description{
Returns the search path for executable files based on
%PATH%
}
\examples{
\dontrun{
system.PATH() #
}
}
\references{
\url{http://www.microsoft.com/resources/documentation/windows/xp/all/proddocs/en-us/ntcmds_shelloverview.mspx?mfr=true}
}

