% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{uninstall.packages}
\alias{uninstall.packages}
\title{uninstalls (removes) Installed Packages}
\usage{
uninstall.packages(pkgs, lib, warning = TRUE, ...)
}
\arguments{
\item{pkgs}{a character vector with the names of the packages to be removed.}

\item{lib}{a character vector giving the library directories to remove the packages from. 
If missing, defaults to the first element in \link{.libPaths}.}

\item{warning}{boolean (TRUE), should a message be printed in various cases.}

\item{...}{currently ignored.}
}
\value{
Invisible NULL
}
\description{
A wrapper for \link{remove.packages}. Useful since it also works if the
package is currently loaded into the workspace.
}
\examples{
\dontrun{
install.packages(c("reshape", "plyr"))
require(plyr)
uninstall.packages(c("reshape", "plyr"))
install.packages(c("reshape", "plyr"))
}
}
\seealso{
\code{\link{install.packages}}, \code{\link{remove.packages}},
\code{\link[installr]{install.packages.zip}}
}
